requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
mediapipe==0.10.9
customtkinter==5.2.2
Pillow
cryptography
